fix: refactor project detector out into analysis
